Converted Pages into modules
