+++ signup.php
@@ -20,7 +20,7 @@ $committee =  getCommittee();
 <link rel="stylesheet" href="https://use.fontawesome.com/releases/v5.7.0/css/all.css" integrity="sha384-lZN37f5QGtY3VHgisS14W3ExzMWZxybE1SJSEsQp9S+oqd12jhcu+A56Ebc1zFSJ" crossorigin="anonymous">
 </head>
 <style>
-<?php include 'css/authenication.css'; ?>
+<?php include 'css/styles.css'; ?>
 </style>
 <body>
 <div class="container-fluid">
@@ -214,7 +214,6 @@ $committee =  getCommittee();
 		</div>
 	</div>
 </div>
-
 <script>
 	<?php include 'js/signup.js' ?>
 </script>
