+++ user/html/adminsidebar.html
@@ -1,0 +1,60 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta http-equiv="X-UA-Compatible" content="IE=edge">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title></title>
+</head>
+<body>
+     <!-- Sidebar -->
+     <ul class="navbar-nav sidebar sidebar-dark accordion" id="accordionSidebar" 
+     style="background:#494263;position: -webkit-sticky;position: sticky;top: 0;height: 100vh;">
+
+        <!-- Sidebar - Brand -->
+        <a class="sidebar-brand d-flex align-items-center justify-content-center" href="">
+          <div class="sidebar-brand-text mx-3">RUIMUN ADMIN</div>
+        </a>
+  
+        <!-- Divider -->
+        <hr class="sidebar-divider my-0" />
+  
+        <!-- Nav Item - Dashboard -->
+        <li class="nav-item">
+          <a class="nav-link" href="dashboard">
+            <i class="fas fa-fw fa-tachometer-alt"></i>
+            <span>Dashboard</span></a>
+        </li>
+        <li class="nav-item">
+          <a class="nav-link" href="members">
+            <i class="fas fa-fw fa-users"></i>
+            <span>Members</span></a>
+        </li>
+        <li class="nav-item">
+          <a class="nav-link" href="payments">
+            <i class="fas fa-fw fa-money-bill-alt"></i>
+            <span>Payments</span></a>
+        </li>
+        <li class="nav-item">
+          <a class="nav-link" href="posts">
+            <i class="fas fa-fw fa-share-square"></i>
+            <span>Posts</span></a>
+        </li>
+        <li class="nav-item">
+          <a class="nav-link" href="docs">
+            <i class="fas fa-fw fa-folder"></i>
+            <span>Materials</span></a>
+        </li>
+  
+        <!-- Divider -->
+        <hr class="sidebar-divider d-none d-md-block">
+  
+        <!-- Sidebar Toggler (Sidebar) -->
+        <div class="text-center d-none d-md-inline">
+          <button class="rounded-circle border-0" id="sidebarToggle"></button>
+        </div>
+  
+      </ul>
+      <!-- End of Sidebar -->
+</body>
+</html>
